refactor camera into container
