+++ app/ios/Quilt/AppDelegate.m
@@ -32,7 +32,7 @@
    */
 
   // jsCodeLocation = [NSURL URLWithString:@"http://localhost:8081/index.ios.bundle?platform=ios&dev=true"];
-  jsCodeLocation = [NSURL URLWithString:@"http://10.6.30.77:8081/index.ios.bundle?platform=ios&dev=true"];
+  jsCodeLocation = [NSURL URLWithString:@"http://10.0.1.198:8081/index.ios.bundle?platform=ios&dev=true"];
 
 
   /**
