+++ app/src/containers/app.js
@@ -5,7 +5,7 @@ import Home from '../components/home';
 import CreateQuilt from '../components/create_quilt';
 // todo: create seperate component/container for quilts
 import ShowQuilts from '../containers/quilts';
-import ShowCamera from '../components/camera';
+import ShowCamera from '../containers/camera';
 import WatchVideo from '../components/video';
 import FriendsContainer from '../containers/friends';
 import NavigationBar from 'react-native-navbar';
